+++ src/controllers/MenuController.php
@@ -59,8 +59,10 @@ class MenuController extends Controller
      *
      * @return mixed
      */
-    public function actionIndex($siteHandle='')
+    public function actionIndex($siteHandle = null)
     {
+        $siteHandle = $siteHandle ?? Craft::$app->getSites()->currentSite->handle;
+
         $objSite = Craft::$app->getSites()->getSiteByHandle($siteHandle);
         if (!$objSite) {
             $siteHandle = Craft::$app->getSites()->getPrimarySite()->handle;
@@ -79,7 +81,7 @@ class MenuController extends Controller
      *
      * @return mixed
      */
-    public function actionMenuNew($siteHandle='')
+    public function actionMenuNew($siteHandle)
     {
         $objSite = Craft::$app->getSites()->getSiteByHandle($siteHandle);
         if (!$objSite) {