+++ src/migrations/m200228_124859_olivemenus_addSiteIdFieldToMenusTable.php
@@ -2,6 +2,7 @@
 
 namespace olivestudio\olivemenus\migrations;
 
+use Craft;
 use craft\db\Migration;
 
 /**
@@ -15,8 +16,6 @@ class m200228_124859_olivemenus_addSiteIdFieldToMenusTable extends Migration
 	 */
 	public function safeUp()
 	{
-		// Place migration code here...
-		
 		if (!$this->db->columnExists('{{%olivemenus}}', 'site_id')) {
 			$this->addColumn('{{%olivemenus}}', 'site_id', $this->integer(11)->after('handle'));
 		}
@@ -27,7 +26,6 @@ class m200228_124859_olivemenus_addSiteIdFieldToMenusTable extends Migration
 	 */
 	public function safeDown()
 	{
-		echo "m200228_124859_olivemenus_addSiteIdFieldToMenusTable cannot be reverted.\n";
-		return false;
+	    $this->dropColumn('{{%olivemenus}}', 'site_id');
 	}
 }