+++ CHANGELOG.md
@@ -54,7 +54,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## 1.1.7 - 2020-02-28
 ### Added multisite option
 
+## 1.1.8 - 2020-05-04
+### Only version bump
+
 ## 1.1.9 - 2020-05-15
-### Migrations are not run when updating the plugin
+### Fixed
 - Migrations are not run when updating the plugin #42
 
+## 1.1.10 - 2020-05-15
+### Fixed 
+- migration/down for multi-site
+- set default site ID for multi-site migrations #44
+- fix URL routing for multi-site (index and create) #43